add db

+++ app/Controllers/DeleteComment.php
@@ -2,6 +2,7 @@
 namespace App\Controllers;
 
 use App\Models\Comment;
+use App\services\Session;
 use App\db\DatabaseConnection;
 
 /**
@@ -21,6 +22,10 @@ class DeleteComment
 
     public function execute(int $identifier)
     {
+        $role = Session::get('role');
+        if ($role !='admin') {
+            throw new \Exception('Page résevée à l\'administration !');
+        }
         $postRepository = new Comment();
         $postRepository->connection = new DatabaseConnection();
         $success = $postRepository->deleteComment($identifier);
